Backend/emails: Move email queuing functions out of __init__.py
