+++ app/backend/src/couchers/email/queuing.py
@@ -1,0 +1,141 @@
+from typing import Any
+
+import yaml
+from sqlalchemy.orm.session import Session
+
+from couchers.config import config
+from couchers.email.queuing import queue_email
+from couchers.i18n import LocalizationContext
+from couchers.jobs.enqueue import queue_job
+from couchers.metrics import emails_counter
+from couchers.proto.internal import jobs_pb2
+from couchers.templating import Jinja2Template, template_folder
+from couchers.utils import now
+
+
+def _queue_email(
+    session: Session,
+    sender_name: str,
+    sender_email: str,
+    recipient: str,
+    subject: str,
+    plain: str,
+    html: str | None,
+    list_unsubscribe_header: str | None,
+    source_data: str | None,
+) -> None:
+    """
+    This indirection is so that this can be easily mocked. Not sure how to do it better :(
+    """
+
+    # Import here to avoid circular dependency
+    from couchers.jobs.handlers import send_email
+
+    payload = jobs_pb2.SendEmailPayload(
+        sender_name=sender_name,
+        sender_email=sender_email,
+        recipient=recipient,
+        subject=subject,
+        plain=plain,
+        html=html,
+        list_unsubscribe_header=list_unsubscribe_header,
+        source_data=source_data,
+    )
+    queue_job(
+        session,
+        job=send_email,
+        payload=payload,
+        priority=5,
+    )
+
+    emails_counter.inc()
+
+
+def queue_email(
+    session: Session,
+    sender_name: str,
+    sender_email: str,
+    recipient: str,
+    subject: str,
+    plain: str,
+    html: str | None,
+    list_unsubscribe_header: str | None = None,
+    source_data: str | None = None,
+) -> None:
+    _queue_email(
+        session=session,
+        sender_name=sender_name,
+        sender_email=sender_email,
+        recipient=recipient,
+        subject=subject,
+        plain=plain,
+        html=html,
+        list_unsubscribe_header=list_unsubscribe_header,
+        source_data=source_data,
+    )
+
+
+def queue_userless_email(
+    session: Session, recipient: str, subject: str, template_name: str, template_args: dict[str, Any]
+) -> None:
+    """
+    This is a simplified version of couchers.notifications.background._send_email_notification
+
+    It's for the few security emails where we don't have a user to email but send directly to an email address.
+    """
+
+    # Not yet localizable
+    loc_context = LocalizationContext.en_utc()
+
+    template_args = {
+        **template_args,
+        "header_subject": subject,
+        "footer_timezone_name": loc_context.localized_timezone,
+        "footer_copyright_year": now().year,
+        "footer_email_is_critical": True,  # Results in no unsubscribe footer.
+    }
+
+    # Format plaintext template
+    plain_tmplt_body = (template_folder / f"{template_name}.txt").read_text()
+    plain_tmplt_footer = (template_folder / "_footer.txt").read_text()
+    plain_tmplt = Jinja2Template(source=plain_tmplt_body + plain_tmplt_footer, html=False)
+    plain = plain_tmplt.render(template_args, loc_context)
+
+    # Format html template
+    html_tmplt = Jinja2Template(
+        source=(template_folder / "generated_html" / f"{template_name}.html").read_text(), html=True
+    )
+    html = html_tmplt.render(template_args, loc_context)
+
+    queue_email(
+        session,
+        sender_name=config["NOTIFICATION_EMAIL_SENDER"],
+        sender_email=config["NOTIFICATION_EMAIL_ADDRESS"],
+        recipient=recipient,
+        subject=config["NOTIFICATION_PREFIX"] + subject,
+        plain=plain,
+        html=html,
+        source_data=config["VERSION"] + f"/{template_name}",
+    )
+
+
+def queue_system_email(session: Session, recipient: str, template_name: str, template_args: dict[str, Any]) -> None:
+    source = (template_folder / "system" / f"{template_name}.md").read_text(encoding="utf8")
+    _, frontmatter_source, text_source = source.split("---", 2)
+
+    loc_context = LocalizationContext.en_utc()
+    rendered_frontmatter = Jinja2Template(source=frontmatter_source, html=False).render(template_args, loc_context)
+    frontmatter = yaml.load(rendered_frontmatter, Loader=yaml.FullLoader)
+
+    plain = Jinja2Template(source=text_source, html=False).render(template_args, loc_context)
+
+    queue_email(
+        session,
+        sender_name=config["NOTIFICATION_EMAIL_SENDER"],
+        sender_email=config["NOTIFICATION_EMAIL_ADDRESS"],
+        recipient=recipient,
+        subject=config["NOTIFICATION_PREFIX"] + frontmatter["subject"],
+        plain=plain,
+        html=None,
+        source_data=template_name,
+    )
