+++ app/backend/src/couchers/notifications/background.py
@@ -10,7 +10,7 @@ from couchers import urls
 from couchers.config import config
 from couchers.context import make_background_user_context
 from couchers.db import session_scope
-from couchers.email import queue_email
+from couchers.email.queuing import queue_email
 from couchers.i18n import LocalizationContext
 from couchers.models import (
     Notification,
